requirements.txt
websocket-client
pyaudio
python-docx
